refactor: skill ut ikke-komponenter fra tsx-filer
Fjerner de fire siste react/only-export-components-advarslene.

BehandlingskoerContext og nyeOgFerdigstilteOppgaverTsType inneholdt ingen
komponent i det hele tatt, bare typer, en enum og en context. Advarselen kom
av at Vite behandler .tsx som en Fast Refresh-grense. Riktig fiks var derfor
å endre filendelse til .ts, ikke å splitte filene. Importene er uendret siden
filendelser utelates.

UttrekkResultatDialog inneholder en faktisk komponent og trengte ekte splitt.
harFormatering og formatCelleVerdi er flyttet til uttrekkFormatering.ts.
Den nye fila importerer utils/dateUtils fordi formatCelleVerdi er avhengig av
dayjs duration-pluginen som registreres der.
