+++ src/client/app/avdelingsleder/lagredeSøk/uttrekk/UttrekkResultatDialog.tsx
@@ -1,5 +1,4 @@
 import React, { useContext, useEffect, useMemo, useState } from 'react';
-import dayjs from 'dayjs';
 import { ChevronDownIcon, EyeIcon } from '@navikt/aksel-icons';
 import { ActionMenu, Alert, BodyShort, Button, Dialog, Loader, Pagination, Table } from '@navikt/ds-react';
 import AppContext from 'app/AppContext';
@@ -13,6 +12,7 @@ import {
 	TolkesSom,
 } from 'filter/filterTsTypes';
 import 'utils/dateUtils';
+import { formatCelleVerdi, harFormatering } from './uttrekkFormatering';
 
 const PAGE_SIZE_OPTIONS = [5, 10, 15, 20, 25, 30, 35, 40];
 
@@ -27,68 +27,6 @@ function calculateDefaultPageSize(): number {
 		if (opt <= rows) best = opt;
 	}
 	return best;
-}
-
-export function harFormatering(feltdef: Oppgavefelt | undefined): boolean {
-	if (!feltdef) return false;
-	if (feltdef.verdiforklaringer && feltdef.verdiforklaringer.length > 0) return true;
-	return [TolkesSom.Boolean, TolkesSom.Timestamp, TolkesSom.Duration].includes(feltdef.tolkes_som);
-}
-
-export function formatCelleVerdi(
-	verdi: string | number | boolean | string[] | null | undefined,
-	feltdef: Oppgavefelt | undefined,
-	formater: boolean,
-): string {
-	if (verdi === null || verdi === undefined || verdi === '') {
-		return '-';
-	}
-
-	if (Array.isArray(verdi)) {
-		if (verdi.length === 0) {
-			return '-';
-		}
-		return verdi.map((v) => formatCelleVerdi(v, feltdef, formater)).join(', ');
-	}
-
-	if (!formater || !feltdef) {
-		return String(verdi);
-	}
-
-	if (feltdef.verdiforklaringer && feltdef.verdiforklaringer.length > 0) {
-		const forklaring = feltdef.verdiforklaringer.find((v) => v.verdi === String(verdi));
-		if (forklaring) {
-			return forklaring.visningsnavn;
-		}
-	}
-
-	if (feltdef.tolkes_som === TolkesSom.Boolean) {
-		const boolVerdi = typeof verdi === 'boolean' ? verdi : String(verdi) === 'true';
-		return boolVerdi ? 'Ja' : 'Nei';
-	}
-
-	if (feltdef.tolkes_som === TolkesSom.Timestamp) {
-		const dato = dayjs(String(verdi));
-		if (dato.isValid()) {
-			return dato.format('D. MMM YYYY, [kl.] HH:mm');
-		}
-	}
-
-	if (feltdef.tolkes_som === TolkesSom.Duration) {
-		const strVerdi = String(verdi);
-		const tall = Number(strVerdi);
-		if (!Number.isNaN(tall)) {
-			return `${Math.floor(tall)}`;
-		}
-		const dur = dayjs.duration(strVerdi);
-		const dager = dur.asDays();
-		if (!Number.isNaN(dager)) {
-			return `${Math.floor(dager)}`;
-		}
-		return strVerdi;
-	}
-
-	return String(verdi);
 }
 
 function finnFeltdefForAggregert(
